test: 重构批量构建、流式构建等测试用例，统一使用 testify 断言库并拆分测试场景
1.  新增 streaming、batch、embedding 构建处理器的完整测试用例，替换并迁移原有分散的测试文件
2.  重构已有的测试文件，将单函数测试拆分为带子测试的分组测试，统一使用 testify/assert 和 testify/require 进行断言
3.  删除冗余重复的旧测试文件，合并复用通用工具函数
4.  统一测试代码的上下文注入、任务 payload 处理等公共逻辑
