+++ adp/vega/vega-backend/server/worker/build_task_common_test.go
@@ -9,56 +9,49 @@ import (
 	"errors"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"go.uber.org/mock/gomock"
 
 	"vega-backend/interfaces"
 	vmock "vega-backend/interfaces/mock"
 )
 
-func TestUpdateResourceIndexNameEmptyOldIndex(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	ra := vmock.NewMockResourceAccess(ctrl)
-	resource := &interfaces.Resource{ID: "r1"}
+func TestUpdateResourceIndexName(t *testing.T) {
+	t.Run("updates empty old index", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		ra := vmock.NewMockResourceAccess(ctrl)
+		resource := &interfaces.Resource{ID: "r1"}
 
-	ra.EXPECT().Update(gomock.Any(), resource).DoAndReturn(func(_ context.Context, got *interfaces.Resource) error {
-		if got.LocalIndexName != "new-index" {
-			t.Fatalf("expected new-index, got %q", got.LocalIndexName)
-		}
-		return nil
+		ra.EXPECT().Update(gomock.Any(), resource).DoAndReturn(func(_ context.Context, got *interfaces.Resource) error {
+			assert.Equal(t, "new-index", got.LocalIndexName)
+			return nil
+		})
+
+		require.NoError(t, updateResourceIndexName(context.Background(), resource, ra, "new-index"))
 	})
 
-	if err := updateResourceIndexName(context.Background(), resource, ra, "new-index"); err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-}
+	t.Run("skips unchanged index", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		ra := vmock.NewMockResourceAccess(ctrl)
+		resource := &interfaces.Resource{ID: "r1", LocalIndexName: "same-index"}
 
-func TestUpdateResourceIndexNameUnchangedIndex(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	ra := vmock.NewMockResourceAccess(ctrl)
-	resource := &interfaces.Resource{ID: "r1", LocalIndexName: "same-index"}
-
-	if err := updateResourceIndexName(context.Background(), resource, ra, "same-index"); err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-}
-
-func TestUpdateResourceIndexNameUpdateFailureKeepsOldIndex(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	ra := vmock.NewMockResourceAccess(ctrl)
-	resource := &interfaces.Resource{ID: "r1", LocalIndexName: "old-index"}
-
-	ra.EXPECT().Update(gomock.Any(), resource).DoAndReturn(func(_ context.Context, got *interfaces.Resource) error {
-		if got.LocalIndexName != "new-index" {
-			t.Fatalf("expected attempted update to new-index, got %q", got.LocalIndexName)
-		}
-		return errors.New("update failed")
+		require.NoError(t, updateResourceIndexName(context.Background(), resource, ra, "same-index"))
 	})
 
-	err := updateResourceIndexName(context.Background(), resource, ra, "new-index")
-	if err == nil {
-		t.Fatal("expected update error")
-	}
-	if resource.LocalIndexName != "old-index" {
-		t.Fatalf("expected in-memory LocalIndexName restored to old-index, got %q", resource.LocalIndexName)
-	}
+	t.Run("keeps old index after update failure", func(t *testing.T) {
+		ctrl := gomock.NewController(t)
+		ra := vmock.NewMockResourceAccess(ctrl)
+		resource := &interfaces.Resource{ID: "r1", LocalIndexName: "old-index"}
+
+		ra.EXPECT().Update(gomock.Any(), resource).DoAndReturn(func(_ context.Context, got *interfaces.Resource) error {
+			assert.Equal(t, "new-index", got.LocalIndexName)
+			return errors.New("update failed")
+		})
+
+		err := updateResourceIndexName(context.Background(), resource, ra, "new-index")
+
+		require.Error(t, err)
+		assert.Equal(t, "old-index", resource.LocalIndexName)
+	})
 }
