+++ adp/vega/vega-backend/server/worker/build_task_reconciler_test.go
@@ -8,6 +8,9 @@ import (
 	"testing"
 	"time"
 
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
 	"vega-backend/interfaces"
 )
 
@@ -16,36 +19,30 @@ import (
 // findStuckBuildTasks 是对账判定核心：init 超过时限且队列中无对应消息的任务判为卡死。
 
 func TestFindStuckBuildTasks(t *testing.T) {
-	now := time.Date(2026, 6, 11, 22, 0, 0, 0, time.UTC)
-	staleAfter := 3 * time.Minute
-	ms := func(t time.Time) int64 { return t.UnixMilli() }
-
-	mkTask := func(id, status string, updatedAgo time.Duration) *interfaces.BuildTask {
-		return &interfaces.BuildTask{ID: id, Status: status, UpdateTime: ms(now.Add(-updatedAgo))}
-	}
-
-	tasks := []*interfaces.BuildTask{
-		mkTask("lost", interfaces.BuildTaskStatusInit, 10*time.Minute),     // 卡死：超时且不在队列
-		mkTask("queued", interfaces.BuildTaskStatusInit, 10*time.Minute),   // 正常排队：队列中有消息
-		mkTask("fresh", interfaces.BuildTaskStatusInit, 10*time.Second),    // 刚创建：创建→入队存在间隙，不能误判
-		mkTask("running", interfaces.BuildTaskStatusRunning, 10*time.Minute), // 防御：非 init 不碰
-	}
-	queued := map[string]struct{}{"queued": {}}
-
-	stuck := findStuckBuildTasks(tasks, queued, now, staleAfter)
-
-	if len(stuck) != 1 || stuck[0].ID != "lost" {
-		ids := []string{}
-		for _, s := range stuck {
-			ids = append(ids, s.ID)
+	t.Run("returns init tasks that are stale and absent from queue", func(t *testing.T) {
+		now := time.Date(2026, 6, 11, 22, 0, 0, 0, time.UTC)
+		staleAfter := 3 * time.Minute
+		ms := func(t time.Time) int64 { return t.UnixMilli() }
+		mkTask := func(id, status string, updatedAgo time.Duration) *interfaces.BuildTask {
+			return &interfaces.BuildTask{ID: id, Status: status, UpdateTime: ms(now.Add(-updatedAgo))}
 		}
-		t.Fatalf("expected exactly [lost], got %v", ids)
-	}
-}
+		tasks := []*interfaces.BuildTask{
+			mkTask("lost", interfaces.BuildTaskStatusInit, 10*time.Minute),
+			mkTask("queued", interfaces.BuildTaskStatusInit, 10*time.Minute),
+			mkTask("fresh", interfaces.BuildTaskStatusInit, 10*time.Second),
+			mkTask("running", interfaces.BuildTaskStatusRunning, 10*time.Minute),
+		}
+		queued := map[string]struct{}{"queued": {}}
 
-func TestFindStuckBuildTasks_EmptyInputs(t *testing.T) {
-	now := time.Now()
-	if got := findStuckBuildTasks(nil, nil, now, time.Minute); len(got) != 0 {
-		t.Fatalf("expected empty result for nil input, got %d", len(got))
-	}
+		stuck := findStuckBuildTasks(tasks, queued, now, staleAfter)
+
+		require.Len(t, stuck, 1)
+		assert.Equal(t, "lost", stuck[0].ID)
+	})
+
+	t.Run("returns empty for nil input", func(t *testing.T) {
+		got := findStuckBuildTasks(nil, nil, time.Now(), time.Minute)
+
+		assert.Empty(t, got)
+	})
 }
